move files to ecs dir
