+++ lib/broadside/ecs/ecs_manager.rb
